+++ examples/14-slot-capabilities/mfs/three.js
@@ -1,0 +1,19 @@
+System.register([], (_export, _context) => {
+  function setup(api) {
+    api.registerComponent('component', {
+      mount(container) {
+        container.innerHTML = '<div>From <b>THREE</b>!</div>';
+      },
+      unmount(container) {
+        container.innerHTML = '';
+      },
+    });
+  }
+
+  _export('setup', setup);
+
+  return {
+    setters: [],
+    execute() {},
+  };
+});