+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Improved released package on JSR (only considering native output)
 - Added more inline documentation / type declarations
+- Added `loading-template-id` to attributes (#8)
+- Added slot ordering attributes to `pi-slot` (#7)
 
 **Note**: ⚠️ Early preview: Not suitable for production yet